refactor: Remove Location code type from getInputCodeType function
The code changes involve removing the "Location" code type from the getInputCodeType function in the code-input component. This code type was commented out and not being used, so it has been removed to improve code clarity and maintainability.

Note: The recent user commits and repository commits have been considered to generate this commit message.